Add component infrastructure.
Can finally connect to a server now! (With a ton of error messages.)

+++ src/greenscripter/minecraft/play/inventory/ItemUtils.java
@@ -53,7 +53,7 @@ public class ItemUtils {
 		Slot s = new Slot();
 		s.present = count > 0;
 		s.itemId = type;
-		s.itemCount = (byte) count;
+		s.itemCount = count;
 		return s;
 	}
